pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "charter-parser"
version = "0.1.0"
description = "Extract structured legal clauses from a voyage charter party PDF using deterministic layout parsing plus LLM-assisted cleanup."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "AI Engineer Candidate" }]
license = { text = "MIT" }
dependencies = [
    "pdfplumber>=0.11.4",
    "pydantic>=2.7",
    "openai>=1.50.0",
    "python-dotenv>=1.0.1",
    "typer>=0.12.5",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
charter-parser = "charter_parser.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF", "N", "PL"]
ignore = ["PLR0913"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
